npcm7xx: modify arch and dts
1. remove unused code
2. remove secure test in poleg.c and use secure_boot.c secure test
3. add soc and cpu struct to dts
4. modify wdt dts compatible name
