--- arch/arm/mach-npcm/npcm7xx/cache.c
@@ -1,21 +1,0 @@
-/*
- * SPDX-License-Identifier:	GPL-2.0+
- */
-
-#include <common.h>
-#include <cpu_func.h>
-
-#ifndef CONFIG_SYS_DCACHE_OFF
-
-void enable_caches(void)
-{
-	dcache_enable();
-}
-
-void disable_caches(void)
-{
-	dcache_disable();
-}
-
-#endif
-
