+++ board/nuvoton/poleg/poleg.c
@@ -72,135 +72,6 @@ static int board_sd_clk_init(const char *name)
 
 	return 0;
 }
-
-static bool is_security_enabled(void)
-{
-	u32 val = readl(FUSTRAP);
-
-	if (val & FUSTRAP_O_SECBOOT) {
-		printf("Security is enabled\n");
-		return true;
-	} else {
-		printf("Security is NOT enabled\n");
-		return false;
-	}
-}
-
-static int check_nist_version(void)
-{
-#if (CONFIG_NIST_VERSION_ADDR != 0)
-	volatile u32 uboot_ver = *(u32*)(UBOOT_RAM_IMAGE_ADDR + HEADER_VERSION_OFFSET);
-	volatile u32 nist_ver = *(u32*)(CONFIG_NIST_VERSION_ADDR);
-
-	if (uboot_ver != nist_ver)
-		reset_cpu(0);
-#endif
-
-	return 0;
-}
-
-static int secure_boot_configuration(void)
-{
-#if defined(CONFIG_SPI_FLASH) && defined(SPI_FLASH_BASE_ADDR)
-
-	const u8 tag[SA_TAG_FLASH_IMAGE_SIZE] = SA_TAG_FLASH_IMAGE_VAL;
-	struct spi_flash *flash;
-	struct udevice *udev;
-	u32 addr, addr_align;
-	int rc , i, offset;
-	u8 *buf = NULL;
-
-	rc = spi_flash_probe_bus_cs(CONFIG_SF_DEFAULT_BUS, CONFIG_SF_DEFAULT_CS,
-			CONFIG_SF_DEFAULT_SPEED, CONFIG_SF_DEFAULT_MODE, &udev);
-	if (rc)
-		return rc;
-
-	flash = dev_get_uclass_priv(udev);
-	if (!flash)
-		return -1;
-
-	// fuse images should be a part of the flash image, right after the uboot
-	addr = POLEG_UBOOT_END;
-
-	// if found, program the image to the fuse arrays, set the secure boot
-	// bit and erase the image from the flash
-	if (((u32*)(addr + SA_TAG_FLASH_IMAGE_OFFSET))[0] == ((u32*)tag)[0] &&
-		((u32*)(addr + SA_TAG_FLASH_IMAGE_OFFSET))[1] == ((u32*)tag)[1]) {
-
-		u8 fuse_arrays[2 * NPCM_OTP_ARR_BYTE_SIZE];
-		u32 fustrap_orig;
-
-		printf("%s(): fuse array image was found on flash in address 0x%x\n", __func__, addr);
-
-		memcpy(fuse_arrays, (u8*)addr, sizeof(fuse_arrays));
-
-		fustrap_orig = *(u32*)(fuse_arrays + SA_FUSE_FLASH_IMAGE_OFFSET);
-
-		//TODO: Here, randomize 4 AES keys + generate their nibble parity + embed to image
-
-		printf("%s(): program fuse key array from address 0x%x\n", __func__, addr + SA_KEYS_FLASH_IMAGE_OFFSET);
-
-		rc = fuse_prog_image(NPCM_KEY_SA, (u32)(fuse_arrays + SA_KEYS_FLASH_IMAGE_OFFSET));
-		if (rc != 0)
-			return rc;
-
-		// clear oSecBoot, will be programmed only after everything is
-		// programmed successfuly
-		fustrap_orig = *(u32*)(addr + SA_FUSE_FLASH_IMAGE_OFFSET);
-		*(u32*)(fuse_arrays + SA_FUSE_FLASH_IMAGE_OFFSET) &= ~FUSTRAP_O_SECBOOT;
-
-		printf("%s(): program fuse strap array from address 0x%x\n", __func__, addr + SA_FUSE_FLASH_IMAGE_OFFSET);
-
-		rc = fuse_prog_image(NPCM_FUSE_SA, (u32)(fuse_arrays + SA_FUSE_FLASH_IMAGE_OFFSET));
-		if (rc != 0)
-			return rc;
-
-		// erase the whole sector
-		addr_align = addr & ~(u32)(flash->erase_size -1);
-
-		offset = addr - addr_align;
-		printf("%s(): erase the sector of addr 0x%x\n", __func__, addr_align);
-
-		addr_align -= SPI_FLASH_BASE_ADDR;
-
-		buf = (u8 *)malloc(flash->erase_size);
-		if (buf) {
-			spi_flash_read(flash, addr_align, flash->erase_size, buf);
-		} else {
-			printf("%s(): failed to alloc buffer, skip otp program\n", __func__);
-			return -1;
-		}
-
-		for (i = 0 ; i < (SA_TAG_FLASH_IMAGE_OFFSET + SA_TAG_FLASH_IMAGE_SIZE) ; i++)
-			buf[offset + i] = 0xff;
-
-		rc = spi_flash_erase(flash, addr_align, flash->erase_size);
-		if (rc != 0)
-			return rc;
-
-		rc = spi_flash_write(flash, addr_align, flash->erase_size, buf);
-		if (rc != 0)
-			return rc;
-
-		free(buf);
-
-		// programm SECBOOT bit if required
-		if (fustrap_orig & FUSTRAP_O_SECBOOT) {
-			printf("%s(): program secure boot bit to FUSTRAP\n", __func__);
-			rc = fuse_program_data(NPCM_FUSE_SA, 0, (u8*)&fustrap_orig, sizeof(fustrap_orig));
-		} else {
-			printf("%s(): secure boot bit is not set in the flash image, secure boot will not be enabled\n", __func__);
-		}
-
-		return rc;
-	}
-	// No fuse image was found in flash, continue with the normal boot flow
-
-#endif
-
-	return 0;
-}
-
 int board_init(void)
 {
 	struct npcm_gcr *gcr = (struct npcm_gcr *)npcm_get_base_gcr();
@@ -213,6 +84,7 @@ int board_init(void)
 	gd->bd->bi_boot_params = (PHYS_SDRAM_1 + 0x100UL);
 
 	nodeoff = -1;
+
 	while ((nodeoff = fdt_node_offset_by_compatible(gd->fdt_blob, nodeoff,
                 "npcm750,runbmc")) >= 0) {
 		/* select DAC2 for VGA output */
@@ -283,7 +155,6 @@ int dram_init(void)
 #ifdef CONFIG_LAST_STAGE_INIT
 int last_stage_init(void)
 {
-	int rc;
 	char value[32];
 	struct udevice *dev = gd->cur_serial_dev;
 
@@ -305,18 +176,7 @@ int last_stage_init(void)
 
 	}
 
-	if (is_security_enabled()) {
-
-		rc = check_nist_version();
-		if (rc != 0)
-			return rc;
-	} else {
-
-		// OTP can be programmed only in basic mode
-		rc = secure_boot_configuration();
-		if (rc != 0)
-			return rc;
-	}
+	arch_preboot_os();
 
 	return 0;
 }
