+++ configs/PolegRunBMC_defconfig
@@ -12,6 +12,7 @@ CONFIG_ENV_SECT_SIZE=0x4000
 CONFIG_ENV_OFFSET=0x100000
 CONFIG_ENV_ADDR=0x80100000
 CONFIG_LAST_STAGE_INIT=y
+CONFIG_SYS_MALLOC_F_LEN=0x1000
 
 CONFIG_DM=y
 CONFIG_DM_SERIAL=y
@@ -113,3 +114,5 @@ CONFIG_WDT=y
 CONFIG_WDT_NPCM=y
 CONFIG_CMD_WDT=y
 # CONFIG_WATCHDOG is not set
+
+CONFIG_NPCM_SECURE_BOOT=y
